Engine: hide internal savegame structs in a separate header
