+++ Engine/game/savegame.h
@@ -15,20 +15,6 @@
 #ifndef __AGS_EE_GAME__SAVEGAME_H
 #define __AGS_EE_GAME__SAVEGAME_H
 
-#if __cplusplus >= 201103L
-#include <memory>
-namespace stdtr1compat = std;
-#else
-#if defined (_MSC_VER)
-#include <memory>
-#elif defined (__GNUC__)
-#include <tr1/memory>
-#endif
-namespace stdtr1compat = std::tr1;
-#endif
-
-#include "ac/common_defines.h"
-#include "media/audio/audiodefines.h"
 #include "main/version.h"
 
 
@@ -114,69 +100,6 @@ struct SavegameDescription
     
     String              UserText;
     ABitmap             UserImage;
-};
-
-// PreservedParams keeps old values of particular gameplay
-// parameters that are saved before the save restoration
-// and either applied or compared to new values after
-// loading save data
-struct PreservedParams
-{
-    // Whether speech and audio packages available
-    int SpeechVOX;
-    int MusicVOX;
-    // Script global data sizes
-    int GlScDataSize;
-    std::vector<int> ScMdDataSize;
-
-    PreservedParams();
-};
-
-// RestoredData keeps certain temporary data to help with
-// the restoration process
-struct RestoredData
-{
-    int                     FPS;
-    // Unserialized bitmaps for dynamic surfaces
-    std::vector<Bitmap*>    DynamicSurfaces;
-    // Scripts global data
-    struct ScriptData
-    {
-        stdtr1compat::shared_ptr<char> Data;
-        size_t              Len;
-
-        ScriptData();
-    };
-    ScriptData              GlobalScript;
-    std::vector<ScriptData> ScriptModules;
-    // Room data (has to be be preserved until room is loaded)
-    Bitmap                 *RoomBkgScene[MAX_BSCENE];
-    short                   RoomLightLevels[MAX_REGIONS];
-    int                     RoomTintLevels[MAX_REGIONS];
-    short                   RoomZoomLevels1[MAX_WALK_AREAS + 1];
-    short                   RoomZoomLevels2[MAX_WALK_AREAS + 1];
-    int                     RoomVolume;
-    // Mouse cursor parameters
-    int                     CursorID;
-    int                     CursorMode;
-    // General audio
-    struct ChannelInfo
-    {
-        int ClipID;
-        int Pos;
-        int Priority;
-        int Repeat;
-        int Vol;
-        int VolAsPercent;
-        int Pan;
-        int PanAsPercent;
-        int Speed;
-    };
-    ChannelInfo             AudioChans[MAX_SOUND_CHANNELS + 1];
-    // Ambient sounds
-    int                     DoAmbient[MAX_SOUND_CHANNELS];
-
-    RestoredData();
 };
